Added a moving trajectory algorithm
